+++ assets/javascripts/MapController.js
@@ -2,7 +2,7 @@ import BrandImageControl from "./BrandImageControl.js";
 import CopyrightControl from "./CopyrightControl.js";
 import LayerControls from "./LayerControls.js";
 import TiltControl from "./TiltControl.js";
-import { capitalizeFirstLetter } from "./utils.js";
+import { capitalizeFirstLetter, preventScroll } from "./utils.js";
 import { getApiToken, getFreshApiToken } from "./osApiToken.js";
 import {defaultPaintOptions} from "./defaultPaintOptions.js";
 
@@ -109,6 +109,7 @@ export default class MapController {
     }
     this.addControls()
     this.addClickHandlers();
+    this.overwriteWheelEventsForControls();
 
     const handleMapMove = () => {
       const center = this.map.getCenter()
@@ -190,6 +191,12 @@ export default class MapController {
       this.layerControlsComponent = new LayerControls(this, this.sourceName, this.layers, this.availableLayers, this.LayerControlOptions);
       this.map.addControl(this.layerControlsComponent, 'top-right');
     }
+  }
+
+  overwriteWheelEventsForControls() {
+    const mapEl = document.getElementById(this.mapId)
+    const mapControlsArray = mapEl.querySelectorAll('.maplibregl-control-container')
+    mapControlsArray.forEach((mapControls) => mapControls.addEventListener('wheel', preventScroll(['.dl-map__side-panel__content']), {passive: false}));
   }
 
   addClickHandlers() {
@@ -440,18 +447,7 @@ export default class MapController {
   createFeaturesPopup(features) {
     const wrapper = document.createElement('div');
     wrapper.classList.add('app-popup');
-    wrapper.onwheel = (e) => {
-      let list = e.target.closest('.app-popup-list');
-
-      if(!list){
-        e.preventDefault();
-        return
-      }
-
-      var verticalScroll = list.scrollHeight > list.clientHeight;
-      if(!verticalScroll)
-        e.preventDefault();
-    };
+    wrapper.onwheel = preventScroll(['.app-popup-list']);
 
     const featureOrFeatures = features.length > 1 ? 'features' : 'feature';
     const heading = document.createElement('h3');